Update default for max_global_series_per_metric limit to 0 (#2714)
This brings the CLI default and Helm chart in line with the Jsonnet
default value. This default is set because it doesn't provide any
protection for Mimir now that series are sharded by all labels
(previously they were sharded only by metric name). It's still useful
for preventing cardinality explosions in some cases but, more often
it is a source of confusion for users.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>