Dashboards: stop setting 15s "interval"
This appears to stem from missing features and confusion: it was
originally set at 1m, and I think that was because Grafana at the time
didn't have `$__range_interval`.

Grafana now lets you configure a scrape interval on your datasource,
which is more likely to be correct for your environment than
hard-coding it into these dashboards.