ruler: ignore non-config values when checking defaults
The idea of `rulestore.IsDefaults()` is to check whether the ruler has
been configured by the user. However `reflect.DeepEqual()` checks all
data within the Config struct, including things which are not actually
config.

This change skips any fields within the config marked as `yaml:"-"`,
which are not settable from yaml and therefore not config.

Using package `github.com/google/go-cmp/cmp`, which is documented as
"not for use in production" since it panics when it cannot compare
things. However we should catch that in unit testing, so should be ok.
(Note gRPC also uses this package).

Added a test which (artificially) induces the problem.