Add the ability to reject object storage requests past a limit
Add an experimental option for the store-gateway to reject requests over
the limit of inflight requests to object storage. This potentially helps
improve latency when one particular store-gateway is overloaded with
expensive object storage requests.

This behavior is optional and defaults to off, preserving the existing
behavior of blocking when the maximum number of inflight requests to
object storage is exceeded.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>