Add a timeout to query sharding process
Adds a context check to expression mapping when running query sharding
or instant query splitting.

Also a timeout of 10s is set on the context passed to the mapping.
We don't expect this operation to last more than few milliseconds, so
anything reaching 10s can be considered a bug or poor implementation.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>