Refactor label value cardinality tests/benchmark
Refactored these tests to use real TSDB instead of a mocked one, so
changes to the code won't need changes in the tests, which have
incomplete or unreal mocks.

Also dropped the LabelValueCardinality_ContextCancellation as that test
wasn't testing any specific test where context cancellation would be
taken into account.

Instead, added `TestCountLabelValueSeries_ContextCancellation` which
tests that specific function, which is the slowest, to respect the
context.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>