Replace internal thanos fork with latest thanos
Due to the out-of-order implementation and Mimir's dependency with
Thanos's code, we had to fork thanos due to an interface change in the
TSDB.

Since out-of-order has been upstreamed to Prometheus and Thanos has
already updated its dependency, we can now get rid of our internal fork
usage.

Please note that Thanos has extracted its objstore implementation to its
own repository so we had to update a few references throughout the
codebase.

Signed-off-by: Jesus Vazquez <jesus.vazquez@grafana.com>