Compactor: emit histogram of block max time delta to now (#3240)
* Compactor: emit histogram of block max time delta to now

Emit a histogram of the difference between `now` and the max time of
a block being compacted as a number of seconds. This can be used to
alert if compactors are not able to keep up with the number of blocks
being generated. In this case, the delta will begin to fall into higher
and higher buckets (> 24h).

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>

* Changelog

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>