Updated detailed steps for migrating blocks from Thanos to Mimir.
The detailed steps on how to migrate the blocks from Thanos to Mimir isn't available and the current documentation also doesn't seems to be helpful. Therefor I have formulated the steps I have followed for migrating the blocks from Thanos to Mimir so that it would be helpful for people who are looking to move from Thanos.