Updated detailed steps for migrating blocks from Thanos to Mimir. (#3290)
* Updated detailed steps for migrating blocks from Thanos to Mimir.

The detailed steps on how to migrate the blocks from Thanos to Mimir isn't available and the current documentation also doesn't seems to be helpful. Therefor I have formulated the steps I have followed for migrating the blocks from Thanos to Mimir so that it would be helpful for people who are looking to move from Thanos.

* Updated the changes suggested after review

* Replaced xargs with for loop as suggested

* Updated the documentation with the changes suggested.

Updated the documentation with the changes suggested with the for loop and also included steps to remove the external labels from the meta.json as well.

* Updated with the changes suggested

* Updated with the changes suggested

Updated the script to remove labels from meta.json, retention of the downsampled blocks and other changes suggested.

* Updated with the changes suggested.

* Updated the changes suggested

* Updated with the changes suggested.

I have updated the PR with the AWS command to copy blocks and corrected the typos.