Do chunks cleanup outside chunksReader
This commit shows how the pooled chunks bytes can be put back into the
pool outside the chunksReader. The main problem that this is solving is
how to make the link between
 * an allocated byte slice from the block's byte pool. This happens
   in the chunk reader after it retrieves the chunk bytes from the
   bucket or from the cache
 * and using and discarding a series in BucketStore.Series()

When BucketStore.Series() iterates it doesn't have a reference from
which block's pool the chunk's bytes were taken from.

This is solved by the chunksKeeper which is used to allocate slices
from the pool and to reclaim them when they are done with. A SeriesSet
implementation (reclaimingChunksSeriesSet) keeps track of when a series
is used.

A possible problem with this approach is that the link is kept based on
the pointer of the chunk. So if the chunk's .Data is copied to a new
storepb.Chunk, then we do not release the pooled bytes until the
Series() method returns.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>