Reduce default value of chunk_pool_min_bucket_size_bytes to 128
In production we see that the chunk bytes pool is returning byte slices
that are too big. Sometimes 200-300 times bigger than requested. This
doesn't have a negative effect because the chunk reader used to use a
single byte slice for many chunks, so it likely used most of the pooled
bytes.

However, for streaming store-gateway it is more difficult to do this
optimization.

The current minimum bucket size is 16K. This commit changes it to 128
so we can use the pool for smaller byte slices without having to
optimize how we use the pooled bytes.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>