store-gateways: extract read locking during Series
Series() acquires a per-tenant lock when starting to read from a block.
The reason is to prevent a race between addBlock, removeBlock, and
reading from these blocks (for example a block is unloaded after Series
() has already selected it for querying; when accessing this block's
index header we will be using an mmaped file which was deleted from
disk).

This PR moves this locking outside of Series() so we can use further
simplify the code in PR 3540 (see [comment](https://github.com/grafana/mimir/pull/3540#discussion_r1033623997))

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>