Tag all activity-tracker logs
It can be quite hard to find all the activity-tracker logs, I usually
grep by `activity` but some customers have that in many queries that are
being logged too. Grepping by `activity=` could do the trick sometimes,
but also not always.

This adds a `module=activity-tracker` label to all the logs issues for
the activity tracker, making it much easier to find them.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>