streaming store-gateway: use an interface for chunkReader
I wanted to test loadingBatchSet. This was difficult because we have no
control over the chunks in the blocks, so it's difficult to assert on
what chunks we get.

With this PR I am making the chunkReader an interface so that it's
easier to inject chunks and check that chunkReaders (plural) and
loadingBatchSet are working correctly together.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>