Streaming store-gateway: add structs and interfaces for sets of series
This is the first PR in which we are introducing code related to the
streaming implementation of the store-gateway.

This PR introduces two prod files and two test files.

- `series_refs.go` contains structs and interfaces for holding and
iterating over a set of series which still doesn't have chunks yet.
It only contains refs to chunks and labels.

- `series_chunks.go`, in turn, contains structs and interfaces for
representing and iterating over a set of series which has its labels
_and_ chunks loaded.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>
Co-authored-by: Marco Pracucci <marco@pracucci.com>
Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>