Parse req form again after setting default time
When setting the default "time" param, if the form was already parsed
(by some previous middleware) we didn't propagate that change into
r.Form.

In order to solve that, we're now replacing the r.Body with a
readCloseSeeker, that can be seeked to the beginning again and again.

This change isn't so important on itself, but it's needed to make
https://github.com/grafana/mimir/pull/3561 work properly, as there we're
going to parse the request form at the beginning of the request in order
to generate the activity.

Additionally, though not important, note that we were never checking the
POST request body, in case the "time" was specified there, and that's
where Prometheus API client is sending it, so we were always setting the
default "time" param on the Query, and it worked just because if it was
also sent in the body, that would take precedence when parsing the form
(but it was a waste of resources since we had to parse and render the
query string again all the times).

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>