Streaming store-gateway: Add tests for blockSeriesChunkRefsSetsIterator
These tests use the e2e framework. The framework is not ideal for these
tests because the blocks it creates have a lot of variability. This
is also the reason why we cannot assert more exactly on the selected
chunk refs.

The next step for this will be to refactor
blockSeriesChunkRefsSetsIterator to be able to circumvent the coupling
that an indexReader has to a BucketStore.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>