Use BasicLifecycler for Compactor and autoforget unhealthy instances
Change the compactor to use BasicLifecycler instead of Lifecycler so that
we can make use of autoforget functionality. This works around an issue
where ownership of tenants is retained by unhealthy instances.

The behavior of compactors while starting changes slightly because of the
use of BasicLifecycler instead of Lifecycler.

* Lifecycler didn't join the ring when starting, only after running so the
  `wait_active_instance_timeout` was only needed while waiting for the instance
  to become active.
* BasicLifecycler joins the ring while starting, thus we need to apply the
  `wait_active_instance_timeout` when starting the lifecycler, _not_ just
  when waiting for the instance to become active.

See #3708
Fixes #1588

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>