ruler: set default for for-grace-period=2m and update docs
This change reflects how we operate the ruler at Grafana Labs.

The benefit is that alerts no longer retrigger when a ruler starts when
their "for" period is longer than the for-grace-period.

A possible drawback of having a reduced grace period means that upon a
restart the ruler will send queries to the ingesters (or long-term
storage, depending on the value of `ruler.for-outage-tolerance`)to query
 the state of alerts. These queries are very light and do not show to be
  an issue.

Related to issue 2888

Fixes issue 3814

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>