Adopt upstream change to allow Iterator reuse
Upstream Prometheus changed `Iterator()` to optionally take an old
iterator for reuse, to avoid allocating a new one and hence reduce
garbage-collection. We need to change Mimir to have this extra
parameter.

Actual reuse of iterators inside Mimir would come in a later change;
mostly the parameter is ignored here.