Fixes templating parsing for manually created vars
When you create a dashboard variable from within the Grafana UI, it
creates the query field as an object. The object contains a refId as
well as the original query.

The existing code expects the query to be a string, not an object
and will therefor miss any metrics used if the variable was created
in the UI.