Vendor update mimir-prometheus to f3d1f7756fcc (#3912)
* Vendor latest mimir-prometheus

For fixes to native histograms and block compaction

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>

* Adopt upstream change to allow Iterator reuse

Upstream Prometheus changed `Iterator()` to optionally take an old
iterator for reuse, to avoid allocating a new one and hence reduce
garbage-collection. We need to change Mimir to have this extra
parameter.

Actual reuse of iterators inside Mimir would come in a later change;
mostly the parameter is ignored here.

* Cope with upstream change to Series interface

Upstream Prometheus changed from taking a `Labels` to be populated to
taking a `ScratchBuilder` which you can call `Labels()` on.

Since `ScratchBuilder.Labels()` returns a newly-allocated slice we don't
need to copy it. It's possible this extra allocation will be significant
and need an additional method to mitigate.

* Cope with upstream changes to relabel.Process

Previously a "drop" command was indicated by zero-length labels return;
now it is indicated by an explicit bool return being false.

* Cope with upstream change to NewAPI

Upstream Prometheus added a `ScrapePoolsRetriever` object; since Mimir
doesn't do any scraping we just dummy it out.

* Cope with upstream change for FloatHistogram

`storage.Appender.AddHistogram` needs to take an extra parameter.
The only place this caused an error was in a dummy method.

* Use replace in go.mod to point to the correct grafana/regexp version

Otherwise go mod tidy takes the newer main version which is not what
we want.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>
Co-authored-by: Bryan Boreham <bjboreham@gmail.com>