Compactor: count compactions errors and context cancelations separatelly
A context cancellation during a scheduled compaction is currently
treated as an error and the `cortex_compactor_runs_failed_total`
and `cortex_compactor_tenants_processing_failed` counters are
incremented.

This can throw off one of the alerts in the mixin when the compactor
restarts more than once every 2 hours.

```yaml
  - alert: MimirCompactorHasNotSuccessfullyRunCompaction
    annotations:
      message: Mimir Compactor {{ $labels.pod }} in {{ $labels.cluster }}/{{ $labels.namespace}} failed to run 2 consecutive compactions.
      runbook_url: https://grafana.com/docs/mimir/latest/operators-guide/mimir-runbooks/#mimircompactorhasnotsuccessfullyruncompaction
    expr: |
      increase(cortex_compactor_runs_failed_total[2h]) >= 2
    labels:
      reason: consecutive-failures
      severity: critical
```

This PR adds a `reason` label on the metric to differentiate between
failures due to shutdowns and failures due to errors.

It also changes to alert to take this label into account. The alert should
still work with compactors that don't emit this label.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>