Alertmanager: Expose missing upstream metrics from the dispatcher. (#4151)
There are two metrics in the Dispatcher component of the upstream Alertmanager
we don't currently expose:

- `alertmanager_dispatcher_aggregation_groups`
- `alertmanager_dispatcher_alert_processing_duration_seconds`

I can't see any reasoning behind ommitting these, so I am assuming it was
perhaps just an oversight, and adding them in this commit. I've chosen not
to expose metrics per-user - we can consider this if a need arises.

The change also re-orders the code so all the Dispatcher metrics are grouped
together, and extends the unit test to get coverage on the existing metric
(`cortex_alertmanager_dispatcher_aggregation_group_limit_reached_total`).