ingester: elide two function calls in pushMetadata
Function `appendMetadata` was only ever called once; eliding it allows
the call to `getOrCreateUserMetadata` to be hoisted out of the loop
and save locking on every series.

The call to `GetMetricFamilyName` checked if the input was nil, which
probably won't happen, but if it did we would rather ignore that input
than store metadata under the metric name `""`.