Querier: batch merge iterator handle histograms
The batch merge iterator is default interator used to let the querier merge
the chunks received from store-gateways.

The chunk.Batch and stream merge code is optimized to the extreme.
Adding histograms into the Batch with any additional indirection destroys
the performance, hence the solution proposed in the PR which has 0 additional
allocation when the code deals with float series.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>