Validate query-ingesters-within limit
Previously, we checked that the query-store-after duration was less than
the old query-ingesters-within, otherwise there might be a gap that we
don't query. Updated the validation for that we check against the new
query-ingesters-within limit instead. Note that only the new default
limit is checked and only on startup, so if there's an override for a
tenant it won't be checked. This is consistent with how other limits are
checked.