Queries dashboard: display store-gateway bytes/sec instead of items/sec
The two panels for data fetched/sec and touched/sec currently show the
number of series, chunks or posting lists that the store-gateway
processes per second. They are also put on the same scale, but a single
posting list can have a lot bigger impact on query performance than a
single chunk.

Usually query performance and store-gateway resource usage are more
closely tied to the number of bytes processed per second rather than the
 number of items processed.

So this PR makes a change to display the number of bytes per sec for
postings, chunks and series (both touched and fetched) instead of the
number of items.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>