Set MetricTarget type to Utilization to align with usage of averageUtilization
When averageUtilization is used as the MetricTarget, Kubernetes expects the type to also be Utilization. Currently, the usage of averageUtilization overrides the averageValue type and K8s defaults to a type of Utilization instead. This causes Helm to show drift, as the template wants averageValue but the applied resource has Utilization.

It is assumed that the intent is for the type to be Utilization given the usage of averageUtilization and the values file referring to percentage targets rather than absolute value targets.