store-gateway index header: return posting offset in LabelVales (#4641)
This PR introduces the LabelValuesOffsets method to indexheader.Reader and indexheader.PostingsOfssetTableV(1|2). The method creates behaves the same was as LabelValues but instead of only the string values of each posting list, it returns the offsets in the index file of each posting list. This method is currently unused, but will be used once I make progress on issue 4593.

For PostingsOfssetTableV1 I chose to copy the implementation of LabelValues since it is short and trivial.

For PostingsOffsetTableV2 I extracted the traversal logic from LabelValues into a new method postingOffsets which takes a transformer, which is called with every matching posting list. The result of the transformer is appended to a slice. Both LabelValues and LabelValuesOffsets use postingOffsets, but pass different transformers. The former accumulates extracts only the string label values, the latter accumulates whole PostingListOffset.

---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>
Co-authored-by: Marco Pracucci <marco@pracucci.com>