store-gateway: add config param for series selection strategy
Series (or postings) selection strategy is a way for the store-gateway
to choose for which matchers to fetch postings and which to apply on
already fetched series. Previous PRs added ways to propagate
and apply these deferred matchers.

This PR is one of the last major pieces of work in issue 4593 - adds
different strategies and a configuration parameter which selects the
strategy. This PR adds 3 more strategies:
* `worstCase` - this strategy optimizes for the worst case in terms of
fetching data. It assumes that the query will fetch all the series from
the smallest posting list. It uses this number to calculate the size of
these series. It selects posting lists such that their total size is
less than the size of the series from the smallest posting list.

* `speculative` - this assumes that each subtracting posting group will
reduce the size of selected series by some fraction. From there it works
in a similar fashion to the `worstCase` strategy

* `fixedFactor` - this strategy doesn't fetch postings lists that are
100x the size of the smallest posting list.

I ran these strategies against a block index and a day's from an
internal GEM cluster in Grafana Labs. I tried tuning the parameters of
two:
* `worstCase`
  * size of each series = 512, size of each posting list = size in block
  <results>
  * size of each posting list - sizes of lists in the cache are smaller
  than the bucket. I ran this strategy assuming each size is 70% smaller
  than in the block (similar compression to the index cache).
   <results>
  * size of each series - tried with an estimation of series size of 512
   and 256
   <results>

* `speculative`
  * assuming each intersecting posting list reduces the series by 50%

    This showed 92% decrease in fetched postings, but also 2.6x increase
    in fetched series.
    <results>

  * assuming each intersecting posting list reduces the series by 25%

    This showed similar results to the worstCase scenario.

  * I decided to go with the 0.5 speculative paramter so we have some
 more variety in live testing

* `fixedFactor`
  * factor of 10: the rate of fetched series was much higher than the
  baseline implementation

  * factor of 100:
  <results>

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>