Add prepare-shutdown endpoint to ingesters for down scaling
This change adds a new HTTP endpoint to ingesters that changes their
in-memory configuration (with an on-disk backup) such that they will:
* Unregister from the ring when they stop
* Flush all in-memory data to object storage when they stop

This differs from the shutdown endpoint because it does not actually
stop the ingesters, just modifies their configuration in preparation
for being permanently stopped. This is a requirement of using the
rollout-operator for gracefully scaling down ingesters in Kubernetes.

See grafana/rollout-operator#47

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>