store-gateway: add config param for series selection strategy (#4698)
What this PR does
Series (or postings) selection strategy is a way for the store-gateway
to choose for which matchers to fetch postings and which to apply on
already fetched series. Previous PRs added ways to propagate
and apply these deferred matchers.

This PR is one of the last major pieces of work in issue 4593 - adds
different strategies and a configuration parameter which selects the
strategy. This PR adds 3 more strategies.

Strategies
worstCase - this strategy optimizes for the worst case in terms of
fetching data. It assumes that the query will fetch all the series from
the smallest posting list. It uses this number to calculate the size of
these series. It selects posting lists such that their total size is
less than the size of the series from the smallest posting list.
This strategy comes in two flavours:

worstCase
worstCaseSmallPostingLists - this uses the actual size of posting lists
as it would be in the index cache instead of the block
speculative - this assumes that each subtracting posting group will
reduce the size of selected series by some fraction. From there it works
in a similar fashion to the worstCase strategy


---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>