store-gateway: improve series filtering metrics
Around the work of issue 4593 we added the possibility to fetch series that don't necessarily match the request.
However, we aren't really measuring the proportion of series that are discarded because they don't match the request.

Ideally we'd want to measure the volume of data that is discarded, but this is more difficult. I believe that measuring
the number of series will be a good proxy for the volume of data too. This PR does that.

This PR adds the `stage` label on `cortex_bucket_store_series_data_touched{data_type="series"}` metrics. It can take
the value `returned` or `processed` (same as the `chunks` equivalent).

This PR also fixes a type in the mixin and clarifies the help texts of `cortex_bucket_store_series_result_series`.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>