Make `query-ingesters-within` per-tenant (#4287)
* Add query-ingesters-within per-tenant override

Added a new QueryIngestersWithin field to the Limits struct. This will
allow a per-tenant value to be set for this field. The field is not
being used by the querier yet, that will come in a later commit. The old
flag has not been deleted yet.

* Use the query-ingesters-within tenant override

This commit starts using the query-ingesters-within tenant override
when deciding whether to query the distributor or not. This requires the
userID to be passed in/extracted from the context in a couple of places
since we need to do a tenant override lookup.

* Calculate shuffle sharding ingester lookback based on override

The shuffle sharding lookback value is calculated based on whether
shuffle sharding is enabled and whether query-ingesters-within is
greater than zero. As we've changed to using a tenant override for
getting query-ingesters-within, the lookback calculation had to be
changed.

Previously, the lookback was calculated once at startup, since
query-ingesters-within was a global, unchanging value. With the move to
tenant overrides that can be changed at runtime, the lookback
calculation has been moved so it's done every time we get the ingesters
for querying.

* Validate query-ingesters-within limit

Previously, we checked that the query-store-after duration was less than
the old query-ingesters-within, otherwise there might be a gap that we
don't query. Updated the validation for that we check against the new
query-ingesters-within limit instead. Note that only the new default
limit is checked and only on startup, so if there's an override for a
tenant it won't be checked. This is consistent with how other limits are
checked.

* Remove old query-ingesters-within flag

* Use same queryIngestersWithin value for all Querier methods

* Preserve querier yaml field

* Use placeholder values for the Mimir version for deprecated flag removal

* Set default limit properly

* Use ingester's TSDB retention for shuffle sharding lookback

We only need to find ingesters for shuffle sharding for the duration
that blocks are retained on the ingesters. query-ingesters-within can be
much longer if the OOO time window is large e.g. a week, a month. Using
query-ingesters-within could mean we're unnecessarily querying some
ingesters.

* Run make reference-help

* Fix distributor tests after move to ingester retention value for lookback

* Update shuffle sharding documentation after lookback change

* Update CHANGELOG

* Fix shuffle sharding integration test

* Only check query bounds once when creating a distributor querier

* Keep newline in querier doc

* Reuse pre-existing lookback field for TSDB retention being passed to distributor

* Use lookbackPeriod variable

* Revert comment after lookback variable change

* Fix modules.go reference

* Test userID for distributor queryable

* Remove outdated TODO

distributor.go is no longer modified by this PR so the TODO to add tests
here is not needed anymore.

* Remove outdated comment

* Add global flag test

* Fix TestBatchMergeChunks

* make docs

* Keep checking that read shuffle sharding is enabled

* Fix reference to tsdb retention

Previous field used was set by tsdbIngesterConfig() rather than in
yaml/command line.

* Improve validation for query-ingesters-within

Previous approach ignored per-tenant limits and only validated against
the default limits when the application was starting up. New approach
validates query-ingesters-within every time a config is loaded at
runtime.

* Update Mimir removal version

* Lint mimir_test.go

* Update Mimir version references after rebase