store-gateway: reduce series size estimation (#4962)
The 64KiB size estimation for series was inherited from Thanos. The PR which adds this to Thanos didn't give
any explanation for why 64KiB was chosen. Based on some rough estimation we think that we can bring this number down
in order to reduce the volume of fetched data.

Citing Marco
> Assuming a worst case scenario of 1 sample /sec, in a 24h block we have 744 chunks per series.
> The worst case scenario of each meta ref is 8*3=24 bytes, so 744*24 = 17856 bytes.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>