store-gateway: record index header loading times separately
Currently, the time to load an index can either be attributed to loading `expand_postings` or to `fetch_series_and_chunks`.
This depends on which stage needed to read the index header first (posting lists vs symbols).
This makes it harder to tell whether there's something in either of those two stages which increases
latency or it's the index header.

This PR forces to load an index header upon opening a `bucketIndexReader`. Then records this time
in a new stage in `cortex_bucket_store_series_request_stage_duration_seconds` - `load_index`.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>