Remove write path utilization based limiting
Considering that block compaction will potentially use a lot of CPU,
it's probably a bad idea to start rejecting write requests when the CPU
usage is high, because it would mean in practice that we prioritize
compaction over writes. This is not desired, rather we want to
prioritize write requests over read requests. As such, limiting read
requests should be enough.

Signed-off-by: Arve Knudsen <arve.knudsen@gmail.com>