Store-gateway: change defaults for postings strategies
This PR enabled postings strategies by default for the 2.9 release.

We have been running the store-gateway with postings strategies at Grafana Labs for some weeks now and the `worst-case`
strategy is giving the best results so far.

The series preference of 0.75 seems to be a reasonable balance for cases where the optimization results in increased pressure to memcached.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>