Store-gateway: disable index header verification
This PR disables verifying the index header upon loading it. Reading the whole index header
can be resource consuming. Verificaiton/checksumming reads the whole index header once in addition to the read
to sample the index header. This change should effectively halve the time to load an index header since one of the
reads will be skipped.

pracucci suggested to make this configurable per-tenant. But the changes would involve changing 23 tests and
3 more production code usages. I'm not sure if the complexity is worth it at this point. I can extend the PR if
reviewers think this change is incomplete.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>