Store-gateway: disable index header verification (#5174)
* Store-gateway: disable index header verification

This PR disables verifying the index header upon loading it. Reading the whole index header
can be resource consuming. Verification/checksumming reads the whole index header once in addition to the read
to sample the index header. This change should effectively halve the time to load an index header since one of the
reads will be skipped.

---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>