compactor: Add timeouts on blocks cleanup
This PR adds 1-minute timeouts for fetching individual files during compactor user cleanup.
* reading meta.json for one block
* reading deletion or no-compact markers for one block

It has the side effect of setting a timeout to reading the bucket index for all consumers (compactor, store-gateway, querier).
Previously the querier was using a 15-second timeout and the others were unbounded. Since this is already a very generous timeout,
I don't expect it to have any negative effects on the other components.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>