Use slice of []mimirpb.Exemplar (not pointers)
This code validates that the elements of provided slice are not nil,
however I couldn't find any references to why the would be nil.

`prompb.WriteRequest` specifies them as structs (not pointers), there
are no tests checking it, and I can't find any code path where they
would be nil.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>