store-gateway: always include `__name__` posting group in selection
Excluding the `__name__` posting group resulted in increased API calls because the selected series were more sparse in
the index. The reason for this is that series are first sorted by the `__name__` label if they don't have any labels
that start with an uppercase letter. Selecting sparse series results in more API calls because the series are more
likely to belong to different metric names and be spread across the TSDB index.

This PR modifies the `worst-case` strategy to always include `__name__`. I opted for including this change in the existing
`worstCaseFetchedDataStrategy` as opposed to wrapping `worstCaseFetchedDataStrategy` in a new strategy because it will
make testing a bit easier and there's no need for modularity because we're eventually aiming to have only a single strategy.
Happy to rework this though.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>