store-gateway: clean up spans
While reviewing 5182 I realized that the spans we have in the store-gateway aren't very consistent and could use some
more information. This PR does this. It leaves out the bucket_store_merge_all span, which is also somewhat incorrect,
but that is being improved in 5182 (and I don't want to create extra conflicts with that PR)

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>