store-gateway: wait for query gate after loading blocks
A surge in lazy loading index headers can saturate the query concurrency limit and effectively block any other queries from even starting.
This was evident in a recent similar incident where the request rate on the full query path was impacted by store-gateways lazy-loading blocks.

Below are screenshots of request rate, latency, lazy loading latency and query gate latency reported by store-gateways.

We already have a concurrency gate for lazy loading index headers added by Heather, which prevents from overloading the disk.
If we decouple the query gate and the header loading gate, queries that require loading many index headers will be waiting
on a separate gate and not impact queries which are querying already loaded blocks.

#### Note to reviewers

This is a draft PR because I suspect it will conflict with the changes in stream-store, so I will rebase it after merging that PR first.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>