Alertmanager: Support changing the default templates for all tenants.
This change adds a new configuration option `-alertmanager.default-template`
which allows the system operator to specify a template file to be parsed
prior to tenant template files. What this essentially provides is a simple
way for operators to alter the default templates (which we inherit from
upstream Alertmanger), without rebuilding Mimir.