Make block upload logging more verbose
We should log as "Info" the operations where block upload is started and
completed, as those are important to know when the blocks were uploaded.

Leaving as debug the individual files logs, as there could be tons of
them.

Also refactored to have a single logger with correct operation and a
feature="block upload" fields, instead of adding the operation on some
log lines (but not other lines).

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>