ruler: leave ring before shutting down rule evaluation
The goal of this PR is to make the ruler more responsive during shutdown and avoid missed rule evaluations.
This PR does it by first leaving the ring and only then proceeding to stop rule evaluation. This is helpful when
rule evaluations take long to gracefully shut down (or when there is a deadlock?)

#### What this PR does

This PR changes the order of shutdown procedures in the ruler. Previously they were
1. Stop rule evaluations
  * this includes waiting for all tenants and all rule groups to finish evaluating (tenants are waited for concurrently, but rule groups within a tenant are stopped serially)
2. Stop ring lifecycler
  * this changes the ruler state to LEAVING
  * when other rulers detect this ring change, they recalculate rule group ownership and start evaluating the rules of the leaving ruler
3. Stop other dependency services; this includes:
  * propagating & processing notifications from other rulers for changed rule groups
  * updating the ring state from the KV store
  * cleaning up & heartbeating gRPC clients to other ruler replicas

After this PR there is an additional step in the beginning:

0. Change ring state to LEAVING
1. Stop rule evaluations
2. Stop ring lifecycler
  * this changes the ruler state to LEAVING, but this is now a noop
3. Stop other dependency services

This change is safe from the perspective of the leaving ruler because the ring is no longer used after `Ruler.running` exits.

What this change makes more likely are duplicate evaluations: while the leaving ruler is shutting down rule evaluations another ruler may pick up a rule group and start evaluating it. It is possible that the rule groups are evaluated in parallel by the two rulers. But duplicate evaluations should be less harmful than no evaluations. A remediation to this risk can be to stop all rule groups for a tenant concurrently [in upstream Prometheus](https://github.com/grafana/mimir/blob/bd6cdc40ee62e20ed8057f5ec3f8f5f7a3d0c2c5/vendor/github.com/prometheus/prometheus/rules/manager.go#L1091-L1093)

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>