makefile: avoid docker pull on most make targets
The current make target fetch-build-image ensures that the build image is present locally after pulling it from DockerHub.

This PR changes the target to first check whether the image isn't present locally before doing `docker pull`. This speeds up the target and allows it to run without an internet connection.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>