Shard `group` aggregator
Some customers want to count their series, and with adaptive metrics on
Grafana Cloud we suggest them to use the `group by ()` operator.
However, that operator wasn't previously sharded, so we hit the chunks
limit when querying large number of series.

This adds support for query-sharding for the group by operator.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>