Count the number of blocks loaded by a store-gateway by duration
Emit metrics about the number of blocks a store-gateway has loaded based on
the duration of the blocks (bucketed into the block durations the compactor
is configured to produce).

A common symptom of the compactor failing is store-gateways loading more 2h
blocks that usual and degrading performance of the read path. Counting the
number of blocks by duration allows us to detect when this happens.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>